Support for snapping memory state to disk
This is useful for loading definitions, snapping the state of the data
segment, and then converting that to binary and then an object file to
append to the binary.

This process seems extremely fragile, though... It took much fiddling
to get this to work. I suspect virtio code is not correctly waiting
for a descriptor because this only started working when I added .S
calls inside of DUMP-MEMORY to slow everything down.